Manage password containing NULLs
This patch enable the bcrypt module to handle passwords containing NUL
bytes. This is needed for passwords hashed by RT ([1]) which simply pass
the password through the SHA256 encoding, without hexifying them, and
then to bcrypt.

The native code has been modified to not use `strlen` for passwords, and
a new parameter has been added to `bcrypt_hashpass` to pass the actual
length.

[1]: https://github.com/bestpractical/rt/blob/552ebebdcd96396b8cb0b9a10d1f54ac9c7b10e0/lib/RT/User.pm#L989-L1020