Fix incorrect python signature of gensalt
It seems that pyo3 does not support specifying a python `bytes` as default
value: the `signature` attribute accepts rust literals but interprets them as
python values; `b"2b"` is interpreted as a rust `[u8, 2]` and not a python
`bytes` object.

To work around this, the `prefix` argument is handled in code, and the python
signature is overridden via pyo3's `text_signature` attribute, which makes the
desired signature visible to python tooling (such as `inspect.signature`).

Fixes #1030